# docs/swagger.yaml
basePath: /v1
definitions:
  api.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  api.SystemResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  model.APModeRequest:
    properties:
      passphrase:
        type: string
      ssid:
        type: string
    type: object
info:
  contact: {}
  description: iot-wifi-setup helps get connected to local wifi network then hand-off
    to your app
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: iot-wifi-setup
  version: "1.0"
paths:
  /aps:
    get:
      consumes:
      - application/json
      description: List all nearby wifi access points
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SystemResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: List all nearby wifi access points
      tags:
      - network
    put:
      consumes:
      - application/json
      description: Start Access Point mode
      parameters:
      - description: The SSID (required) and passphrase (optional) to use with the
          AP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.APModeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SystemResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Start Access Point mode
      tags:
      - network
  /network:
    put:
      consumes:
      - application/json
      description: Sets the client wifi network connection
      parameters:
      - description: The SSID (required) and passphrase (optional) to use with the
          AP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.APModeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SystemResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Sets the client wifi network connection
      tags:
      - network
  /status:
    get:
      consumes:
      - application/json
      description: Gets current network status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SystemResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Gets current network status
      tags:
      - network
swagger: "2.0"